os/FileStore: make temp object writes conditional on replay
The general pattern of concern is:

1- create temp/foo
2- write temp/foo
3- write temp/foo
4- move temp/foo real/foo
5- restart
6- clear temp/*
7- replay 3 (write temp/foo)
8- replay 4 (move tmp/foo real/foo)

...because it would move a half-written object into position.

Previously, we would avoid this implicitly because the per-pg temp
collection would be created on demand during each peering interval,
and deleted on startup before journal replay.  That means that we would
only succeed *any* temp writes if we started back from the early
interval and had the whole thing.

Now, the temp collections are different because they always exist (or
rather, don't exist at all, and are manufactured by the FileStore
layer to avoid an ugly refactor).

To prevent the same pattern now we prevent any write to a temp
object that doesn't yet exist, except touch, which is always the first
operation during recovery.

We re-use the conditional replay value here.  Almost everyone checks
for > 0 so returning 0 will fail.  Fix touch so that it will proceed,
and add an explicit check in collection_move_rename that the source
exists (this is the only other conditional replay user).

Signed-off-by: Sage Weil <sage@redhat.com>